$small-font-size
This is the smaller captions size font. Currently used nowhere in
bootstrap, it is used extensively throughout out our mockups for things
like captions and legal definitions.

Yet another type size to add to our scale.  This gives us a range of
small to large copy sizes to vary the layouts within the bootstrap
system