Fixes secondary buttons, button groups and split buttons from not having a distinct separating border
This update removed most of button color and background customizations
previously allowed in the variable definitions.

Going forward it appears that the theme requires its users to match
button-type to defined theme-color type. Ex. ‘primary’ button will
always be the defined ‘primary site colour’ unless overridden in a
helper file such as this commit does.

Should not be a problem for us overall, but interesting to see
bootstrap removing a layer of customization previously available by
default.